added search duration for limit mode, sepaarate ui for debt and search
