+++ docs/UserGuide.adoc
@@ -59,7 +59,7 @@ image::Ui/Shortcuts.png[width="500", align="center"]
 Love eating pizza everyday but find it a chore to retype the same thing to keep track of your spending? Add a shortcut so you need to type less to add frequent expenses or incomes!
 
 === Reminders
-Have Dolla let you know when an expense is coming up, and when a budget or saving is about to be exceeded!
+Have Dolla let you know when the due date of your debts is coming soon!
 
 === Settings `[coming in v2.0]`
 Customise various features of Dolla, like how many items to show per page on a list, what kind of things Dolla will remind you about, or even the statistics that's pinned on the app!
